Correcciones y estilos para la pantalla de cines cercanos.
